Implemented gateway functionality
This PR introduces the ability to run the machineserver in gateway mode, where it only accepts portal connections.

This allows a split between the agent handlers, which needs a persisted connection, and the console/portal handlers which require fewer and more short-lived connections.

With gateway mode it is possible to scale both the agent handers and the console handlers.

This update also simplifies several utility functions and fixes some validation issues.